+++ src/node/media_stream_destination.rs
@@ -69,11 +69,11 @@ impl AudioNode for MediaStreamAudioDestinationNode {
         &self.channel_config
     }
 
-    fn number_of_inputs(&self) -> u32 {
+    fn number_of_inputs(&self) -> usize {
         1
     }
 
-    fn number_of_outputs(&self) -> u32 {
+    fn number_of_outputs(&self) -> usize {
         0
     }
 }
